reset 5th feb

+++ src/main/java/com/pemig/api/util/Const.java
@@ -17,7 +17,7 @@ public final class Const {
   public static final String ALL_LOANS = "all_loans";
   public static final String ALL_CARDS = "all_cards";
   /** Tune this to affect how long the JWT token lasts. Default is 5 * 60 * 60, for 5 hours. */
-  public static final long JWT_VALIDITY = 5 * 60 * 60; // 5 hours
+  public static final long JWT_VALIDITY = (long)5 * 60 * 60; // 5 hours
   // Some constants for pagination and sorting
   public static final String DEFAULT_PAGE_IDX = "0";
   public static final String DEFAULT_PAGE_SIZE = "5";
